ts: remove hash cache
the reason i am removing the hash cache is that it means the
transaction is mutable, which creates problems in rust. i will
need to cache the hashes some other way. probably what i will
do is pass in some sort of HashCache object and save the hashes
there. that way we still solve the quadratic hashing issue
without having to make the transaction mutable.