begin earthbucks-node
Currently, the idea is that the library (earthbucks-lib, earthbucks_lib) should
just be a library and should not handle any IO, such as database or HTTP.
Instead, I will put all the node code, such as blockchain database and HTTP
REST API, in a new project called "node". In typescript, it is earthbucks-node
and in rust it is earthbucks_node.

What follows is the planned architecture from the README as it stands today:

All protocol code is implemented in Rust (rs) and TypeScript (rs)
simultaneously with common test vectors.

rs/ts:

- lib (transactions, blocks, data structures, algorithms)
- node (full/spv, merkler, mempool, p2p web API, wallets, mysql)
- pow (CUDA, WebGL, WebGPU, CPU)
- sdk (client)

ts only:
- gui (auth, wallet, explorer, mining pool, button)

Primary full nodes:
- earthbucks.com: mining pool (rs full node + gui)
- impstack.com: mining pool (rs full node + gui)
- satored.com: mining pool (rs full node + gui)

Additional nodes:
- buttonbucks.com: mining pool (rs full node + gui)
- impulsecash.com: mining pool (ts full node + gui)
- compubutton.com: spv wallet (ts spv node + gui)
- openspv.com: spv wallet (rs spv node + gui)