rename variables. work on schema. include tx/block buf.
MySQL can store up to 4 gigabytes in a BLOB. While it is not really optimized
for large data storage, it seems dramatically easier at the start if we simply
store whole transactions and whole blocks in the database. When Bitcoin
began, 1 MB sized blocks were sufficient from about 2009 - 2016, meaning
seven years of 1 MB blocks. That is a sign that even with some modest success,
storing all blocks and txs in the database may not become an issue for a very
long time. We can also prune this data after the blocks/txs are verified and
use only the indexed form.