new architecture. use node.js for db/api for now.
I'm just trying to build the software in the most effective way possible, and I
find the tools for dealing with my database are better in node.js (Drizzle)
than in rust (sqlx). And of course I know from experience node.js is great for
networking applications like the API. So I will just plan to have all IO in
node.js except for the stuff necessary to make the builder work. Most likely,
all IO will end up being implmented in both languages. But not at this moment.