finding blocks works.
as of this running, it finds up to 6 blocks. theoretically, even though PoW is
nearly nonexistent (the only thing that changes is the timestamp), it should
actually keep finding blocks and settle out to a 10 minute block interval with
time.

a lot of things are still missing. i plan for the node to validate its own
blocks. there also needs to be a broadcast and voting mechamism. and of course
support for adding new transactions.

but the basics of building a blockchain are now there! a blockchain is being
built!