lock_time -> lock_block_num
simpler to just use block_num, with no option to use time. this also makes it
very clear that we can include the current block_num into the coinbase
transaction, which both causes each coinbase tx to be unique, even with the
same inputs/outputs, and also makes the logic of putting this into block
verification easier to understand.