block verification works (on coinbase tx only).
block verification is now working and does in fact validate the blocks that are
being produced. in this moment, i am at block 8 of a new blockchain. however,
the blockchain only includes coinbase transactions at this time. i will need to
produce a wallet to be able to add new transactions to the blockchain to know
if it is really working.