earthbucks-com is not open source
What I will do is extract all open-source software from
earthbucks-com and place it into separate locations such as
earthbucks-pool and earthbucks-wallet when the time is right. For
now I plan to build the app in a proprietary way before extracting
the open-source, since I think that will be easier.