easy deployment with fly.io
earthbucks-com is the new app which i actually plan to host on AWS using ECS.
however, i am more familiar with fly.io, so i have made it work there first.
theoretically, the same container that runs on fly.io will also run on ECS, so
gettin it to work on fly.io first is a nice check.

earthbucks-web is still in this repo and that refers to the landing page that
is hosted at earthbucks.com. the names are confusing. eventually,
earthbucks-web will go away and it will be replaced with earthbucks-com.