matmul 1024
this works, however, it turns out the hashes are the bottleneck. it takes 77
milliseconds to perform the 1024x1024 matrix times 1024 vector. but 64 ms of
that is just computing the hashes. thus, this won't work, because we are still
bottlenecked by hashes.

next, i'm going to explore using a standardized psuedorandom number generator
to see if this goes faster. my goal is to make this bottnecked by matrix
multiplication, not by hashes.