progress. 670 ms vs 1 ms.
it takes 1.75ms to produce a 256x256 matrix of bits from 256 blake3 hashes on
the source data. and then it takes 670 ms to produce the square of that matrix.
this is what i've been looking for. producing the matrix with a hash takes far
longer than the matmul.

Time elapsed in create_256x256_matrix_of_bits is: 1.75ms
Time elapsed in matrix.dot(&matrix) is: 630.960667ms