currently running with 256x256 version
this takes ~2 ms to create the matrix with serial blake3 hashes

but it takes ~670ms to square the matrix with a GPU-friendly matrix
multiplication

next i need to test this on GPU, but my theory is that this will be far faster
on GPU, while still not being limited by hashes. we shall see.

basically, if when running on a modern GPU, the matrix multiplication is still
slower than hashing, that's what we want. while at the same time, it shouldn't
be ridiculously slow (like the 1024x1024 matrix which takes 40 seconds),
because we still need to validate this on CPU if possible.