matmul in earthbucks-tf
note how it is slower for 256, but faster for 400 and bigger. this is not yet
optimized for node.js. i think the reason the first one is slower is probably
tensorflow getting set up for the first time.

earthbucks-tf (node.js):
✓ matmul256 (186 ms)
✓ matmul400 (115 ms)
✓ matmul512 (219 ms)
✓ matmul1024 (1873 ms)

earthbucks-lib (node.js):
✓ matmul256 (89 ms)
✓ matmul400 (270 ms)
✓ matmul512 (573 ms)
✓ matmul1024 (3623 ms)

rust:
Time for matmul-256: 8.062708ms
Time for matmul-400: 26.306792ms
Time for matmul-512: 85.821917ms
Time for matmul-1024: 689.571625ms