creating matrix test
create 1024 matrix: 49.157958984375 ms
matmul1024:         79.796142578125 ms

This is the test I needed to perform. On my machine, creating a matrix (by
doing a lot of blake3 hashes) is still a significant percentage of the time it
takes to perform the matmul. What this means is that we need to make the matrix
multiplication even harder to make sure it is a larger fraction of the
computational speed.

We can do this in two ways: Increase the difficulty of the matrix
multiplication (bigger matrix, or a cube instead of a square). Or, decrease the
difficulty of the hash function, by reducing the number of hashes (say, by
using a smaller matrix, but then also increasing the matrix from a square to a
cube).