copying hash is a lot faster
takeaway: reducing the number of hashes to create the matrix makes a big
difference. even though i am using wasm for blake3 in the browser, it still
saves a significant amount of time by not making that many hashes.

the alternate strategy is to just copy the first hash. subsequent hashes
theoretically provide no additional pseudorandomness, so why not just use a
single hash? not only is this faster, but this may allow me to use the pure js
version of the hash, saving a lot of hassle getting blake3 in the browser.

by using a single hash, the time to produce a large matrix is cut down to about
1/10. instead, most of the remaining time is just the time it takes to send the
matrix to the GPU.

takeaway: we should copy the first hash to save CPU.

replicating hash over and over:

create 2048 matrix arr: 18.278076171875 ms
create 2048 matrix: 54.05810546875 ms
matcube2048: 209.045166015625 ms

hashing the hash over and over:

create 2048 matrix arr: 141.740966796875 ms
create 2048 matrix: 162.233154296875 ms
matcube2048: 313.6162109375 ms