rename: int1289 and float1289
the theory here is that both of these should be deterministic, and therefore
float1289 is better because it also exploits the GPU's ability to do floating
points. however, i want to keep the integer one around for a while because i
want to do more testing to be absolutely sure about the determinism of both of
these. i will somehow have to test on other software/hardware configurations.