more testing and an insight about entropy
hashing is not as fast as i'd like in a browser, and i really want to make sure
the computation is GPU-bound. i've realized there is a very simple way to get
lots of pseudorandom data into the matrix calculation, which is by including th
previous block headers.

with a 1289x1289 sized matrix, we have 1661521 elements to fill. assuming we
fill one bit per element, at 256 per block ID we need the previous 6490.3 block
Ids. that's a little more than the previous six weeks worth of block IDs. that
should be just fine to load into memory.

also consider that this can be loaded into the GPU once per block. you will
load in the previous six weeks worth of block IDs, plus the "working block ID"
which will change every time the mine adds a new tx into the block header, and
for that matter every time they update the timestamp. so you will need to
update the "working block ID" very frequently, but the other 6000 block IDs
only need to be updated once every ten minutes. that means the amount of data
you need to send to the GPU is not that high. it is just one hash per ten
minutes.

the code needs to be rewritten to reflect this.

my current thinking is:

- load in lots of pseudorandom data from the previous 6000 block headers
- load in one "working block ID"
- do a large integer matrix multiplication
- do a large floating point matrix multiplication with rounding
- reduce the output
- hash the reduction
- compute the new block ID (with "working block ID" inside) .... this is your
  PoW.