remove xor. replace with parallelizeable hash functions.
i've removed the xor because i'm just not convinced it's secure. instead, i've
made the four reductions into four separate hashes. this means they could be
performed in parallel. even just one web worker would be enough to offload the
hashing, but we could use four instead, and that would make the hashing much
faster. although this is slower than the xor, it is definitely more secure.