square 1627
this is a prime number whose square should take about the same number of
operations as the cube of 1289. indeed, it takes about as long to compute on my
machine. (3327ms for 100 operations).

the advantage of using a square is that we are nowhere near reaching a limit
and we will be able to increase the size of the matrix to increase difficulty.
the 1289 cannot be increased in size at all without changing it from a cube to
a square. but the 1627 can be increased in size without changing the
fundamental algorithm at all.

thus we can call this new algorith "prime square", where the advantage is that
we can pick any larger prime to increase the difficulty. we can thus gradually
increase the difficulty with time if we choose. or even automatically increase
the difficulty if a certain threshold is reached.