Uint8Array -> Buffer. and fix key test vectors.
The npm module we're using, buffer, is a good way to unify both Uint8Arrays and
node's Buffer class. We can just use "buffer" and forget about all the other
ways of dealing with buffers in node. We should totally eliminate any direct
uses of Uint8Array in the library.