work_algo, work_ser, work_par
work algo: what work algorithm are we using? the idea is that this will change
regularly.

work_ser: the output of a serial work algorithm, such as hashing a hash 500
times.

work_par: the output of a parallel work algorithm, such as a 1000x1000 matrix
multiplication

the pseudorandom data for both work algorithms is the current "working block
id", which is the block id with both work_ser an work_par as all null values.
additionally, block IDs from the block adjustment interval may also be
inclduded, which is useful to create a lot of data for the matrix
multiplication.