switch to pure javascript by default in browser
This is the best solution I have found so far. WASM has all sorts of issues,
most significantly it cannot be loaded synchronously. The asynchronous loading
either doesn't work or just creates a problem when you have to wait for it to
load. For now, I'll just use the pure javascript version in a browser by
default. There is still a separate package available that does WASM in a worker
for those times when we really need maximum speed for tons of hashes, e.g.
hashing a big tx, or doing CPU POW.