autoComplete="off"
i'm not sure whether to enable this or disable it. personally, i find it
annoying. but i'm not sure if some people use the in-browser password manager
and it is actually a requirement for them. i will disable for now and will be
on the lookout if people find that more annoying than the alternative.