domain first
we're going to be able to make all API methods available to the outside. you
must pass in the domain when calling these methods. by adding the domain into
fetch, we're going to be able to call to third party domains. this may not be
used inside the app, but it is what will enable third-party apps to use the
same API.