untested methods to handle session data
sessions will store an ID as a cookie. server-side, we will store/retrieve the
user's public key in the session data.

the "token ID" stored in the datatabase is actually the hash of the "session
ID" stored in the cookie. this prevents a database breach from compromising the
session IDs.