Isomorphic Buffer
And ismorphic hex and isomorphic strings.

After investigating protobufs and gRPC, I think they are great tools for many
use-cases, but the top reason why I can't use them for EarthBucks is that
protobufs are not designed to represent binary encoded data isomorphically.
That is to say, I need the ability to decode binary data into a data structure,
and then encode that same data back into binary, and expect the two binary
structures to be identical. Protobufs to have have this feature.

My code needs a bit of fixing up around edge cases, but it does fundamentally
solve this problem. The name "isomorphic buffer" doesn't seem to be taken, at
least not for anything like this, so I will be using that name.

In fact, I currently plan to change the way my RPC API works to actually use
IsoBufs.