also use try/catch pattern
I'm trying to get my typescript errors to look like rust to make it easier to
manage the two different types of code. Using Result is close, but you simply
cannot propagate errors using a ? operator in typescript like you can in rust.
This is the closest I have come so far. Wrap the internals in a try/catch, and
then use .mapErr just like you would in rust. By then calling .unwrap, it
throws an error if there is an error. You then catch this in the outer
try/catch and simply return a value.

This approach has the advantage that it will also wrap any methods inside in a
try/catch and handle those errors as well. That means any method with a Result
type that uses the internal try/catch will NOT throw and will instead return
the error in the Result enum, exactly like rust.

This requires more code, because it means every method that implements Result
must also include an internal try/catch, but it looks better than any other
method I've seen, and it keeps the behavior as close to Rust as I've been able
to manage so far.