Resultify PermissionToken. no more "throw new Error(...)"
This finishes the Resultification of the typescript code. Nowhere in the main
library do we ever use "throw new Error". Now I need to update the rust code to
have all the same error checking.

Moving forward, the goal is to stick with rust-style Result <Ok,Err> values for
all methods that may have errors.