sequence -> lock_rel, locktime -> lock_abs
The goal of EarthBucks is to enable on-chain transactions in the range of 1 US
cent to 5 US dollars. Payment channels are not useful for this. Furthermore,
since the launch of Bitcoin, payment channels have never been proven useful in
the real world. Instead, the sequence number in Bitcoin has been changed from a
payment channel "sequence number", to a relative lock time. This has been
proven more useful. I have decided to adopt this new meaning of "sequence
number" in EarthBucks.

Therefore, we have:

- lock_abs: The absolute block number after which a transaction is valid. (This
  value is always interpreted as an absolute block number and NOT a time).
- lock_rel: This is what used to be the "sequence number". It is a relative
  lock number in each input that tells you the input is not valid if the output
  was not at least that many blocks earlier.

And to be clear, I DO NOT plan to support payment channels in any form. If we
ever decide they are necessary, we may have to simply create a new transaction
type (e.g., transaction with version byte set to 2).