TxOutBnMap
In order to fully handle CHECKLOCKRELVERIFY (the earthbucks equivalent of
CHECKSEQUENCEVERIFY), we need to know the block number associated with each
output. For that, I have created a new struct called TxOutBnMap (transaction
output block number map), which maps (tx_id,out_num) to the (tx_out,block_num).
We need this in the tx verifier and block verifier, but not in building a tx or
signing a tx.