OP_0 should push empty buffer
for some reason I had previously had OP_0 pushing 0x00 to the stack, but AFAICT
the original bitcoin actually pushed an empty buffer here. it seems more
logical to follow the bitcoin approach of pushing an empty buffer. this also
means that the scriptNum zero is actually also an empty buffer. OP_FALSE
doesn't exist in our language, but if it did, it would also be an empty buffer.