remove pkh2wx from standard
this removes the "two week expiry" output type from standard. the reason for
this is that i think 3 month expiry will work for everyone. i wanted to be able
to pay for an upgrade, but i think a better way to do that is to add a
different type of output altogether: 3 month expiry plus recovery. i can add
another output type later where there is a second recovery key that can be used
after the first one expires. that enables recovery by a mine or wallet if the
user lost their key. that seems worth paying for. i can either add this other
output type before or after launch. adding new output types can be done at any
time.