always use tx_out_bn
in order to build, sign, verify txs, we area always going to be using expiry
txs, and that means we always need the block number (assuming it exists) of the
earlier outputs. if they don't exist, we assume they are unconfirmed, and
therefore have the block number of the current block.