verifyInputLockRel and fix >=
lock_rel is valid if:

self.block_num >= prev_block_num + lock_rel

this is the simplest way to do the calcualation, because if you simply add
lock_rel to the block number you get the new block number where the lock_rel is
valid. also, according to GitHub Copilot, this is the same way it is done in
Bitcoin.