untested new script types. builder does std only.
I'm adding new script types to the tx builder, and I decided to change the
behavior. Instead of auto-determining whether to spend expired outputs or not,
it will always assume you want to spend the output using the default type. This
is the logical way to do it for the tx builder, because the whole point of the
tx builder is to choose from your UTXO set a sufficient number of inputs.
Wheras if you are spending an expired output, that is because you are a mine or
a wallet, and you already know exactly which output you want to spend. In that
case, you want to add a specific input into the builder. For that there is a
new add_input method.

Next, I need to test the new script types for recovery.