ts: tx builder only builds "normal" transactions
as with the rust code, i have now changed the typescript tx builder to only
build "normal" transactions, a.k.a. the pkh portion, but not the expiry or
recovery portion. the reason for this is that the tx builder is designed to
gather sufficient inputs to pay for all the outputs, which is basically the
typical payment flow. that flow is not how to build special inputs as in
recovery or expiry. the way to build those is to insert the input you want
precisely into the tx.

i have also added new recovery script types to the typescript code, but they
have not yet been tested.