jest tests pass
after much effort, i have jest tests passing with these mysterious updates to
the jest config file.

i have this stack overflow Q&A to thank:

https://stackoverflow.com/questions/73735202/typescript-jest-imports-with-js-extension-cause-error-cannot-find-module

now all ts files must have .js in them in order to work correctly in the build.
and i have gotten the jest tests to understand this. this means i am using tsc
for the build and it is building plain js files which work with es module
syntax.

this was far harder than i would have ever expected, but earthbucks-lib is now
AFAICT a normal library that can be imported from the outside.