fix cargo.toml. add new block periods.
The "expiry period" is the length of time UTXOs remain spendable. All UTXOs
expire after 90 days and will immediately be spent by whoever mines that block.

The "expiry safety period" is the period that the mine retains data. The spent
outputs are not deleted after 90 days. They are deleted after 180 days. This
leaves a margin of safety in case there is some sort of recoverable error
pertaining to the expiry period.

After the expiry safetey period, the spent outputs will be deleted. Note that
this is a total of 180 days, but is only 90 days after the end of the expire
period.

Users will need to retain their own transaction history. Mines do not retain
the full transaction history for all time. Archival nodes may archive all data,
but there is no requirement on the network for archival nodes to exist. There
is no guarantee expired data will be retained beyond the expiry safety period.