separate work_par_algo and work_ser_algo
Break apart the "algo" so you can easily specify two different PoW algorithms:
one for serial, and one for paralle. Both are combined in the block header,
enabling fine-tuned PoW for consumer devices. The intention is that both PoW
algorithms will change regularly and unpredictably to discourage the
development of ASICs.

- work_ser: This is designed to hold a serial PoW algorithm, such as serial
  hashes, i.e. blake3(blake3(blake3(...)))
- work_par: This is designed to hold a parallel PoW algorithm, such as a giant
  matrix multiplication. Currently the best candidate is algo1627 held in
  another repo, to be brought here soon.