Revert "move all rust tests to integration tests"
This reverts commit 7dcf673fc91c3e626d520722cbe315a7c82daa9e.

---

I wanted to put all the rust tests in the "tests" folder, similar to
javascript/node.js/typescript. Unfortunately, this causes all the tests to be
built separately and then they run more slowly.

unit :       1.5s
integration: 4.5s

So I've decided to keep them as unit tests in the same file, which is the
idiomatic way to do it in Rust. This means they'll just have to look different
than the TypeScript tests.