"Buffer" -> "EbxBuffer"
i made some progress on making a separate IsoBuf class, but at this moment i've
decided it's too complicated and too different from "Buffer". the only real
problem with "Buffer", the npm module, is that it shares exactly the same name
as the node.js standard library, and it is very easy to accidentally forget to
include "buffer" at the top of the package, meaning i'm mixing the two
different types of "Buffer" which share exactly the same name. A solution is to
create my own Buffer class, called "EbxBuffer", which makes it far easier to
tell if i'm using the correct type. My buffer class is currently the same as
npm module Buffer, except with a different name, intentionally designed to make
it easy to spot usage of the wrong class.