remove unused iso buf packages
after experimenting with a new package for IsoBuf, i've decided the simplest
solution is to leave the rust code a lone, and make a very small change to the
typescript code, where we rename the npm package "Buffer" to "IsoBuf". this
will make it a lot easier to spot any unintentional uses of the node global
Buffer, which we should never use. meanwhile, the only other use case for the
iso buf packages was for strict hex, and i've decided to just keep doing it the
same way i am already doing it with a very simple strict hex module.