StrictHex class in TypeScript. StrictHex trait in Rust.
after some effort, I am simply unable to find an adequate replacement for rust
traits in typescript. the only option i can find is to change the prototype of
the global buffer object. this does not seem like a good idea as it would
impact anybody that wants to use this library. i'd rather leave buffer alone,
which is currently the npm "buffer" package, which is actually a very
thoughtful bridge of node's "Buffer" global and Uint8Arrays. so for now i have
just accepted that there is no way to to do traits in typescript, and the best
i can do is to just use a class. so trait in rust will likely have a
corresponding class in typescript. this means more code in typescript, but that
is already the case.