use bigint
the reason to use bigint always is that they have different behaviors when
dividing by zero. with numbers, division by zero redurns Infinity. but with
bigints, they throw an exception. i want to unify behavior. speed is not the
concern - type safety is. so it is fine to use bigints.