unify header.ts with header.rs ... ts not tested yet
the two header modules are now unified. this is an important step towards
unified block verification. although i don't plan to run a mine with the
typescript code, it is helpful to make sure everything works the same in both
languages as a means of testing. i plan to proceed with the entire blockchain
protocol working evenly across both languages.

now that the two headers are unified, i can proceed with block verification.
most of that code is written already as it relies on transaction verification
and the block builder.

what i plan to do next is basically finish some tools to test and verify an
entire blockchain at high speed. the library will contain all data structures
and algorithms, including PoW, but only with simplistic versions of the
protocol that are OK for testing on a CPU. i will then spool up in-memory
versions of the block builder and have them create and store blocks in memory.
i will also produce transactions this way. this will allow me to build an
entire blockchain for testing things across languages.