always use "SysBuf" in node
to make sure we never use the node global object Buffer. we are always using
the npm package "buffer" which has an equivalent API but which also plays
nicely in the browser. our convention is to always call this "SysBuf" so that
we can notice if we ever accidentally use "Buffer" anywhere.