begin work with tensorflow session
tensorflow in rust is significantly more complicated than js because i have to manage memory myself. so i've decided to completely rewrite this module vs. typescript.

pow_session will do this:

manage an entire proof-of-work (pow) session. this means you load it up with
data (block IDs), and then change that data (working block ID), and for each
change, that is one PoW calculation. the session is many iterations of the same
PoW, that way most of the data stays the same on the GPU, and you are just
getting the results (~5kb or so of data returned from GPU per calculation).