delete earthbucks_pow. update README.
earthbucks_pow is no longer the direction i am going in for
proof-of-work calculations or verifications. we are using tensorflow in
node.js. if we ever need this in rust, it will be delivered via an api.
the computation itself is so much slower than any API call that it is
not a big deal to have a completely separate node.js API service do the
computation or verification.

i have also updated the readme to specify all the projects that are
currently open-source. more tools will come with the goal to run an
entire open-source wallet yourself.